Refactor: ProgramDashboard and CategoryBreakdown charts, improved UX, percent bar, and bugfixes
